+++ examples/PTcell/netParams.py
@@ -32,6 +32,10 @@ netParams.defaultThreshold = 0.0 # spike threshold, 10 mV is NetCon default, low
 netParams.defaultDelay = 2.0 # default conn delay (ms)
 netParams.propVelocity = 500.0 # propagation velocity (um/ms)
 netParams.probLambda = 100.0  # length constant (lambda) for connection probability decay (um)
+netParams.sizeX = 100
+netParams.sizeY = 1350  # cortical depth (will be converted to negative values)
+netParams.sizeZ = 100
+
 
 #------------------------------------------------------------------------------
 # Cell parameters