+++ netpyne/network.py
@@ -1106,6 +1106,24 @@ class Network (object):
         if sim.rank == 0 and sim.cfg.timing: print('  Done; stims modification time = %0.2f s.' % sim.timingData['modifyStimsTime'])
 
 
+    ###############################################################################
+    ### Calculate segment coordinates from 3d point coordinates 
+    ###############################################################################
+    def calcSegCoords(self):   
+        # Calculate relative seg coords for 1 cell per pop, 
+        for pop in self.pops.values():
+            pop.calcRelativeSegCoords()
 
+        # Calculate abs seg coords for all cells
+        for cell in self.cells:
+            cell.calcAbsSegCoords()
 
+    ###############################################################################
+    ### Add 3D points to sections with simplified geometry
+    ###############################################################################
+    def defineCellShapes(self):
+        import sim
+        h.define_shape()
+        for cell in sim.net.cells:
+            cell.updateShape()